Optimised GUI package

+++ src/main/java/com/karuslabs/commons/animation/particles/effect/Context.java
@@ -51,8 +51,8 @@ public interface Context extends Repeatable {
         
     public @Shared Vector getVector();
     
-    public int count();
+    public double count();
     
-    public void count(int count);
+    public void count(double count);
     
 }
